% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haat.R
\name{haat}
\alias{haat}
\title{Calculate Heat Area Above Threshold (HAAT)}
\usage{
haat(datetime, exposure, threshold, steps = 100)
}
\arguments{
\item{datetime}{POSIXct vector, corresponding to the exposure measurement timestamps.}

\item{exposure}{Exposure measurements.}

\item{threshold}{Numeric threshold to be considered, in the same unit as the exposure.}

\item{steps}{interpolation data points for each day, default to 100}
}
\value{
Cumulative daily exposure above the referred threshold (numeric).
}
\description{
Computes the daily cumulative excess exposure above a threshold from
an hourly time series.
}
\examples{
data(example_hourly)

haat(
  datetime = example_hourly$datetime,
  exposure = example_hourly$temp,
  threshold = 32
)

}
